requirements.txt
asgiref==3.5.0
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-cleanup==6.0.0
django-extensions==3.1.5
django-filter==21.1
django-htmx==1.9.0
django-import-export==2.7.1
django-js-asset==2.0.0
django-markdownx==3.0.1
django-mptt==0.13.4
django-mptt-admin==2.2.0
django-multi-form-view==2.0.1
django-sites==0.11
django-tinymce==3.4.0
django-weasyprint==2.1.0
et-xmlfile==1.1.0
fonttools==4.32.0
html5lib==1.1
Markdown==3.3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
pycparser==2.21
pydyf==0.1.2
pyphen==0.12.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tinycss2==1.1.1
tzdata==2021.5
weasyprint==54.3
webencodings==0.5.1
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.1